requirements.txt
feedparser
flask
pymongo
celery
redis
aiohttp
python-dotenv

